requirements.txt
# Core
Django==4.1.6
django-configurations==2.4
gunicorn==20.1.0

# Database
psycopg2-binary==2.9.5
dj-database-url==1.2.0

# Model Tools
django-model-utils==4.3.1

# Rest apis
djangorestframework==3.14.0
dj-rest-auth==2.2.7
dj-rest-auth[with_social]==2.2.7
djangorestframework-jsonapi==6.0.0
djangorestframework-jsonapi[django-filter]==6.0.0
drf_spectacular==0.25.1

# Static and Media Storage
django-storages==1.13.2

# Environments
python-dotenv==0.21.0
